modified price to decimel in DTO + made search using pagination options case-insensitive

+++ src/DTO/ArtworkDTO.cs
@@ -27,7 +27,7 @@ namespace Backend_Teamwork.src.DTO
             public string Title { get; set; }
             public string Description { get; set; }
             public int Quantity { get; set; }
-            public double Price { get; set; }
+            public decimal Price { get; set; }
             public DateTime CreatedAt { get; set; }
             public Category Category { get; set; }
             public User User { get; set; }
@@ -39,7 +39,7 @@ namespace Backend_Teamwork.src.DTO
             public string? Title { get; set; }
             public string? Description { get; set; }
             public int? Quantity { get; set; }
-            public double? Price { get; set; }
+            public decimal? Price { get; set; }
         }
     }
 }
